backend/requirements.txt
# Core application dependencies
fastapi==0.104.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
redis[hiredis]==5.0.0
supabase==2.0.0
openai==1.3.0
tiktoken==0.5.1
langdetect==1.0.9
httpx==0.24.0
spacy
pinecone-client==4.1.1
psycopg2-binary
feedparser
rich>=13.0.0
beautifulsoup4==4.12.2
aiogram==3.3.0
celery==5.3.4
prometheus-client==0.17.1
sentry-sdk==1.32.0
pydantic>=2.4.2
dnspython>=2.3.0

# New dependencies for enhanced functionality
apscheduler==3.10.4
deepl==1.17.0
psutil==5.9.8

# New fetcher dependencies (excluding heavy ML dependencies)
redis[hiredis]>=5.0.0

# Development and testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
Faker
ruff
mypy
types-redis
types-psycopg2
pre-commit 